requirements.txt
bidict==0.23.1
blinker==1.8.2
Brotli @ file:///croot/brotli-split_1714483155106/work
certifi @ file:///croot/certifi_1725551672989/work/certifi
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click==8.1.7
contourpy==1.1.1
cycler==0.12.1
dnspython==2.6.1
filelock @ file:///croot/filelock_1700591183607/work
Flask==3.0.3
Flask-SocketIO==5.4.1
fonttools==4.54.1
fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
gmpy2 @ file:///tmp/build/80754af9/gmpy2_1645455532332/work
h11==0.14.0
idna @ file:///croot/idna_1714398848350/work
importlib_metadata==8.5.0
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2 @ file:///croot/jinja2_1716993405101/work
kiwisolver==1.4.7
MarkupSafe @ file:///croot/markupsafe_1704205993651/work
matplotlib==3.7.5
mkl-fft @ file:///croot/mkl_fft_1695058164594/work
mkl-random @ file:///croot/mkl_random_1695059800811/work
mkl-service==2.4.0
mpmath @ file:///croot/mpmath_1690848262763/work
networkx @ file:///croot/networkx_1690561992265/work
numpy @ file:///work/mkl/numpy_and_numpy_base_1682953417311/work
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.1
pandas==2.0.3
pillow @ file:///croot/pillow_1721059439630/work
psutil==6.1.0
py-cpuinfo==9.0.0
pymongo==4.10.1
pyparsing==3.1.4
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
python-dateutil==2.9.0.post0
python-engineio==4.8.0
python-socketio==5.11.0
pytz==2024.2
PyYAML @ file:///croot/pyyaml_1728657952215/work
requests @ file:///croot/requests_1721410876868/work
scipy==1.10.1
seaborn==0.13.2
simple-websocket==1.1.0
six==1.16.0
smmap==5.0.1
sympy @ file:///croot/sympy_1724938189289/work
thop==0.1.1.post2209072238
torch==2.4.1
torchaudio==2.4.1
torchvision==0.20.0
tqdm==4.66.6
triton==3.0.0
typing_extensions @ file:///croot/typing_extensions_1715268824938/work
tzdata==2024.2
ultralytics==8.3.24
ultralytics-thop==2.0.9
urllib3 @ file:///croot/urllib3_1727769808118/work
Werkzeug==3.0.6
wsproto==1.2.0
zipp==3.20.2
